requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
google-cloud-firestore>=2.11.0
google-cloud-logging>=3.6.0
line-bot-sdk>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
# Google ADK dependencies (assuming standard installation)
google-generativeai>=0.3.0
